Refactor NavigationBar
