+++ front/components/navigation/config.ts
