+++ front/pages/w/[wId]/builder/data-sources/[name]/tables/upsert.tsx
@@ -21,9 +21,9 @@ import { useRouter } from "next/router";
 import { useContext, useEffect, useRef, useState } from "react";
 import { useSWRConfig } from "swr";
 
+import { subNavigationBuild } from "@app/components/navigation/config";
 import AppLayout from "@app/components/sparkle/AppLayout";
 import { AppLayoutSimpleSaveCancelTitle } from "@app/components/sparkle/AppLayoutTitle";
-import { subNavigationBuild } from "@app/components/sparkle/navigation";
 import { SendNotificationsContext } from "@app/components/sparkle/Notification";
 import { guessDelimiter } from "@app/lib/api/csv";
 import { getDataSource } from "@app/lib/api/data_sources";
